refactor: remove deprecated alwaysShow prop
