+++ src/util/IfOverflow.ts
@@ -1,0 +1,1 @@
+export type IfOverflow = 'hidden' | 'visible' | 'discard' | 'extendDomain';
