+++ storybook/stories/API/cartesian/ReferenceArea.stories.tsx
@@ -26,8 +26,8 @@ const StyleProps: Args = {
   ...GeneralStyle,
   ...ReferenceComponentStyle,
   label: {
-    description: `If set a string or a number, default label will be drawn, and the option is content. 
-    If set a React element, the option is the custom react element of drawing label. If set a function, 
+    description: `If set a string or a number, default label will be drawn, and the option is content.
+    If set a React element, the option is the custom react element of drawing label. If set a function,
     the function will be called to render customized label.`,
     table: {
       type: {
@@ -55,30 +55,30 @@ const StyleProps: Args = {
 const GeneralProps: Args = {
   ...ReferenceComponentGeneralArgs,
   x1: {
-    description: `A boundary value of the area. If the specified x-axis is a number axis, the type of x 
-      must be Number. If the specified x-axis is a category axis, the value of x must be one of the 
-      categories. If x1 is not set, the first value on the x-axis is used instead. If one of x1 or x2 
+    description: `A boundary value of the area. If the specified x-axis is a number axis, the type of x
+      must be Number. If the specified x-axis is a category axis, the value of x must be one of the
+      categories. If x1 is not set, the first value on the x-axis is used instead. If one of x1 or x2
       is invalid, the area will not be drawn.`,
     table: { type: { summary: 'string | number' }, category: 'General' },
   },
   x2: {
-    description: `A boundary value of the area. If the specified x-axis is a number axis, the type of x 
-    must be Number. If the specified x-axis is a category axis, the value of x must be one of the 
-    categories. If x2 is not set, the last value on the x-axis is used instead. If one of x1 or x2 is 
+    description: `A boundary value of the area. If the specified x-axis is a number axis, the type of x
+    must be Number. If the specified x-axis is a category axis, the value of x must be one of the
+    categories. If x2 is not set, the last value on the x-axis is used instead. If one of x1 or x2 is
     invalid, the area will not be drawn.`,
     table: { type: { summary: 'string | number' }, category: 'General' },
   },
   y1: {
-    description: `A boundary value of the area. If the specified y-axis is a number axis, the type of y 
-    must be Number. If the specified y-axis is a category axis, the value of y must be one of the 
-    categories. If y1 is not set, the first value on the y-axis is used instead. If one of y1 or y2 is 
+    description: `A boundary value of the area. If the specified y-axis is a number axis, the type of y
+    must be Number. If the specified y-axis is a category axis, the value of y must be one of the
+    categories. If y1 is not set, the first value on the y-axis is used instead. If one of y1 or y2 is
     invalid, the area will not be drawn.`,
     table: { type: { summary: 'string | number' }, category: 'General' },
   },
   y2: {
-    description: `A boundary value of the area. If the specified y-axis is a number axis, the type of y 
-    must be Number. If the specified y-axis is a category axis, the value of y must be one of the 
-    categories. If y2 is not set, the last value on the y-axis is used instead. If one of y1 or y2 is 
+    description: `A boundary value of the area. If the specified y-axis is a number axis, the type of y
+    must be Number. If the specified y-axis is a category axis, the value of y must be one of the
+    categories. If y2 is not set, the last value on the y-axis is used instead. If one of y1 or y2 is
     invalid, the area will not be drawn.`,
     table: { type: { summary: 'string | number' }, category: 'General' },
   },
@@ -103,13 +103,6 @@ const referenceAreaArgTypes = {
   ...InternalProps,
   // Rectangle
   radius,
-  // Deprecated
-  alwaysShow: {
-    description: "Use 'ifOverflow' instead.",
-    table: { category: 'Deprecated' },
-    hide: true,
-    disable: true,
-  },
   // Event handlers
   onClick,
   onMouseDown,
