+++ storybook/stories/API/cartesian/ReferenceLine.stories.tsx
@@ -89,14 +89,6 @@ export default {
     ...InternalProps,
     // Deprecated
     dangerouslySetInnerHTML: { table: { category: 'Deprecated' }, hide: true, disable: true },
-    alwaysShow: {
-      description: "Use 'ifOverflow' instead.",
-      control: { type: 'boolean' },
-      table: { category: 'Deprecated' },
-      hide: true,
-      disable: true,
-      default: false,
-    },
   },
   component: ReferenceLine,
 };
